some renaming, and moving of SCPIBase,Transport and DIO to new directory lib
