+++ pycqed/instrument_drivers/physical_instruments/QuTech/QWG.py
@@ -17,10 +17,10 @@ import os
 import numpy as np
 import logging
 import json
-from typing import Tuple, List, Sequence, Dict
+from typing import List, Sequence, Dict
 
 from .QWGCore import QWGCore
-from pycqed.instrument_drivers.physical_instruments.Transport import Transport
+from pycqed.instrument_drivers.lib.Transport import Transport
 
 from qcodes.instrument.base import Instrument
 from qcodes.instrument.parameter import Parameter
