+++ pycqed/instrument_drivers/physical_instruments/QuTech_AWG_Module.py
@@ -26,7 +26,7 @@ import re
 import json
 from typing import Tuple, List, Sequence, Dict
 
-from pycqed.instrument_drivers.meta_instrument.DIOCalibration import DIOCalibration
+from pycqed.instrument_drivers.lib.DIO import CalInterface
 
 from qcodes.instrument.base import Instrument
 from qcodes.instrument.parameter import ManualParameter
@@ -129,7 +129,7 @@ cw_protocols_mt = {
 # class
 ##########################################################################
 
-class QuTech_AWG_Module(SCPI, DIOCalibration):
+class QuTech_AWG_Module(SCPI, CalInterface):
     __doc__ = f"""
     Driver for a Qutech AWG Module (QWG) instrument. Will establish a connection to a module via ethernet.
     :param name: Name of the instrument  
@@ -314,7 +314,7 @@ class QuTech_AWG_Module(SCPI, DIOCalibration):
         return info
 
     ##########################################################################
-    # overrides for DIOCalibration interface
+    # overrides for CalInterface interface
     ##########################################################################
 
     def calibrate_dio_protocol(self, dio_mask: int, expected_sequence: List, port: int=0):
