+++ pycqed/instrument_drivers/physical_instruments/QuTech_QCC.py
@@ -22,7 +22,7 @@ from typing import Tuple,List
 from .SCPI import SCPI
 from ._QCC.QCCMicrocode import QCCMicrocode
 import pycqed
-from pycqed.instrument_drivers.meta_instrument.DIOCalibration import DIOCalibration
+import pycqed.instrument_drivers.lib.DIO as DIO
 
 from qcodes.instrument.base import Instrument
 from qcodes.instrument.parameter import ManualParameter
@@ -53,7 +53,7 @@ CHAR_MIN = -128
 MAX_NUM_INSN = 2**15
 
 
-class QCC(SCPI, DIOCalibration):
+class QCC(SCPI, DIO.CalInterface):
     """
     This is class is used to serve as the driver between the user and the
     QCC hardware. The class starts by querying the hardware via the
@@ -638,7 +638,7 @@ class QCC(SCPI, DIOCalibration):
         return expected_sequence
 
     ##########################################################################
-    # overrides for DIOCalibration interface
+    # overrides for CalInterface interface
     ##########################################################################
 
     def output_dio_calibration_data(self, dio_mode: str, port: int=0) -> Tuple[int, List]:
