+++ pycqed/tests/dev_qubit_objs/test_qubit_objects.py
@@ -25,7 +25,7 @@ import pycqed.instrument_drivers.physical_instruments.ZurichInstruments.ZI_HDAWG
 from pycqed.instrument_drivers.physical_instruments.QuTech_CCL import dummy_CCL
 from pycqed.instrument_drivers.physical_instruments.QuTech_VSM_Module import Dummy_QuTechVSMModule
 from pycqed.instrument_drivers.physical_instruments.QuTechCC import QuTechCC
-from pycqed.instrument_drivers.physical_instruments.Transport import DummyTransport
+from pycqed.instrument_drivers.lib.Transport import DummyTransport
 
 from qcodes import station
