+++ pycqed/tests/instrument_drivers/physical_instruments/test_QuTechCC.py
@@ -4,7 +4,7 @@ import tempfile
 import os
 from pathlib import Path
 
-from pycqed.instrument_drivers.physical_instruments.Transport import FileTransport
+from pycqed.instrument_drivers.lib.Transport import FileTransport
 from pycqed.instrument_drivers.physical_instruments.QuTechCC import QuTechCC
